refactor: separate UI and logic from `TenantInvitationDrawer`
